backend/requirements.txt
numpy<2
joblib
scipy
scikit-surprise
pandas